requirements.txt
langchain
langchain-openai
openai
python-dotenv
streamlit
pydantic
pandas